Extract NDJSON stream cache
